+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Not released
 
+- onStateChange callback for range widget [#890](https://github.com/CartoDB/carto-react/pull/890)
 - fix LegendWidget update logic for multiple legends case [#889](https://github.com/CartoDB/carto-react/pull/889)
 - onStateChange callback for all widgets [#886](https://github.com/CartoDB/carto-react/pull/886)